Require "host" environment to include a `fetch` API
The primary motivation here is to make the `@data-eden/*` packages just
assume a global `fetch` is available. All browser platforms that we care
about have it already present and Node 18+ has it available by default
(Node >= 16.15.0 can opt-in via `node --experimental-fetch`).

This is a breaking change, but is quite easy for consumers to work
around if they need to support Node < 16.15.0 (or < 18 if they don't
want to use a command line flag). For example, running under Node 14 can
be done by using `cross-fetch` as a dev dependency and requiring
`cross-fetch/polyfill`.

Co-authored-by: David J. Hamilton <dhamilton@linkedin.com>