Update `vite` and `vitest` to latest
Vite does not currently support downstreams with node16 module
resolution (see vite issue 11552).

This causes `tsc` build issues unless we enable `skipLibCheck`.

This means we are no longer testing against the risk of shipping
a type-invalid dependency graph.  We can fix this with a CI job that
builds `tsc` with only dependencies (not dev deps) installed.

Co-authored-by: David J. Hamilton <dhamilton@linkedin.com>