Ensure that `fetch` is only accessed lazily (allow mocking)
Currently, it is not possible to mock `fetch` **after** `buildFetch` is
invoked. This might _seem_ fine, but ultimately it is not:

- Generally speaking we expect `buildFetch` to be done **very** early in
  the application's evaluation. Specifically, this will be **_well_**
  before any individual tests are kicked off / ran.
- Tools like `Pretender` / `MirageJS` / `msw` need to be able to
  override `globalThis.fetch` lazily. Most of the time this is done
  within a specific test context (or after the tests have kicked off, in
  a `beforeEach`/`beforeAll` kinda of situation). Notably, these tests
  will almost certainly be directly `import { fetch } from
  'app-name/network'` or something (which **guarantees** that the
  `buildFetch` will have been called already).

This commit ensures that `globalThis.fetch` is looked up lazily on
`globalThis` each time it is needed (therefore allowing other tools to
mock it as desired).

Note that we still eagerly error of `typeof fetch` is undefined when
calling `buildFetch` as this helps folks absorb the breaking changes from
0.2 to 0.3 (no longer using cross-fetch internally).