chore: remove Fetch type export
We don't want to export a public custom `Fetch` type.  The mental model
users should have is that `buildFetch()` returns a `typeof fetch` --
i.e. a fully fetch-compatible function.

The docs show `type Fetch = typeof fetch` as an example of something
users might do to make their own code more readable, but it is not
something that should be part of `@data-eden/network`'s public API.

Co-authored-by: Robert Jackson <rjackson@linkedin.com>