Ensure that TrackingInfoPerFetch is always shared across all versions.
In some bundling contexts (e.g. webpack as used by ember-auto-import) you might
end up with multiple copies of this module in the final bundle. In that case
we need to ensure that we have a single global tracking map. We do this by
using a symbol that is the same across all copies, and then using that symbol
to store the map on the global object.

This ensures that at any given time (regardless of how many copies of
this module exist) we can always get the same map (and therefore our
API's work properly).

Co-authored-by: David J. Hamilton <dhamilton@linkedin.com>